Added handling of package input, new logic to handle invalid values for package info and unit tests, change package info data types

+++ EverestPackageDeliveryCLI/Models/PackageInfo.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct PackageInfo {
   let packageID: String
-  let packageWeightInKg: Int
-  let distanceInKm: Int
+  let packageWeightInKg: Double
+  let distanceInKm: Double
   let offerCode: String?
 }
